adding task  js in  module
